# Skip Hash#rehash steps when it is comparing by id